refactor(web): move utils/validation to @safe-global/utils
