+++ apps/web/src/components/common/AddressBookInput/index.tsx
@@ -9,7 +9,7 @@ import InfoIcon from '@/public/images/notifications/info.svg'
 import EntryDialog from '@/components/address-book/EntryDialog'
 import css from './styles.module.css'
 import inputCss from '@/styles/inputs.module.css'
-import { isValidAddress } from '@/utils/validation'
+import { isValidAddress } from '@safe-global/utils/utils/validation'
 import { sameAddress } from '@safe-global/utils/utils/addresses'
 
 const abFilterOptions = createFilterOptions({
