+++ apps/web/src/components/safe-apps/AddCustomAppModal/index.tsx
@@ -16,7 +16,7 @@ import CheckIcon from '@mui/icons-material/Check'
 import type { SafeAppData } from '@safe-global/safe-gateway-typescript-sdk'
 import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined'
 import ModalDialog from '@/components/common/ModalDialog'
-import { isValidURL } from '@/utils/validation'
+import { isValidURL } from '@safe-global/utils/utils/validation'
 import { useCurrentChain } from '@/hooks/useChains'
 import useAsync from '@/hooks/useAsync'
 import useDebounce from '@/hooks/useDebounce'
