+++ apps/web/src/components/transactions/TxFilterForm/index.tsx
@@ -13,7 +13,7 @@ import { useMemo, type ReactElement } from 'react'
 
 import AddressBookInput from '@/components/common/AddressBookInput'
 import DatePickerInput from '@/components/common/DatePickerInput'
-import { validateAmount } from '@/utils/validation'
+import { validateAmount } from '@safe-global/utils/utils/validation'
 import { trackEvent } from '@/services/analytics'
 import { TX_LIST_EVENTS } from '@/services/analytics/events/txList'
 import { txFilter, useTxFilter, TxFilterType, type TxFilter } from '@/utils/tx-history-filter'
