+++ apps/web/src/components/tx-flow/flows/AddOwner/ChooseOwner.tsx
@@ -19,7 +19,7 @@ import AddressBookInput from '@/components/common/AddressBookInput'
 import NameInput from '@/components/common/NameInput'
 import { useAddressResolver } from '@/hooks/useAddressResolver'
 import useSafeInfo from '@/hooks/useSafeInfo'
-import { uniqueAddress, addressIsNotCurrentSafe } from '@/utils/validation'
+import { uniqueAddress, addressIsNotCurrentSafe } from '@safe-global/utils/utils/validation'
 import type { AddOwnerFlowProps } from '.'
 import type { ReplaceOwnerFlowProps } from '../ReplaceOwner'
 import TxCard from '../../common/TxCard'
