+++ apps/web/src/components/tx/AdvancedParams/AdvancedParamsForm.tsx
@@ -2,7 +2,7 @@ import { type SyntheticEvent } from 'react'
 import { Button, DialogActions, FormControl, Grid, Typography, DialogContent } from '@mui/material'
 import { FormProvider, useForm } from 'react-hook-form'
 import { safeFormatUnits, safeParseUnits } from '@safe-global/utils/utils/formatters'
-import { FLOAT_REGEX } from '@/utils/validation'
+import { FLOAT_REGEX } from '@safe-global/utils/utils/validation'
 import ModalDialog from '@/components/common/ModalDialog'
 import { AdvancedField, type AdvancedParameters } from './types'
 import GasLimitInput from './GasLimitInput'
